+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ### 0.9.1-pre
 * Add file_path arg to webconsole#install_bundle to allow custom bundle file location
+* Add response code 201 handling as webconsole#install_bundle success 
 
 ### 0.9.0
 * Initial version